refactor: move kubeconfig command
